+++ include/RKCKernel.h
@@ -4,7 +4,7 @@ typedef struct ker_args{
   int id;
   vector *v0;
   int count_v0;
-  double h;
+  float h;
   int n_x;
   int n_y;
   int n_z;
@@ -12,5 +12,5 @@ typedef struct ker_args{
   RungeKuttaBenchmark::Fiber **fibers;
 } kernel_args;
 
-void rk2_caller(vector *v0, int count_v0, double h, int n_x, int n_y, int n_z, vector_field field, RungeKuttaBenchmark::Fiber ***fibers);
-void rk4_caller(vector *v0, int count_v0, double h, int n_x, int n_y, int n_z, vector_field field, RungeKuttaBenchmark::Fiber ***fibers);
+void rk2_caller(vector *v0, int count_v0, float h, int n_x, int n_y, int n_z, vector_field field, RungeKuttaBenchmark::Fiber ***fibers);
+void rk4_caller(vector *v0, int count_v0, float h, int n_x, int n_y, int n_z, vector_field field, RungeKuttaBenchmark::Fiber ***fibers);
